\name{h_t_rh}
\alias{h_t_rh}
\title{h_t_rh}
\usage{
h_t_rh(t, rh, p_atm = 101325)
}
\arguments{
  \item{t}{numeric: temperaure (°C)}

  \item{rh}{numeric: relative humidity (1)}

  \item{p_atm}{numeric: atmospheric pressure (Pa)}
}
\value{
numeric: enthalpy of dry air (kJ/kg)
}
\description{
The reference conditions, where enthalpy is 0 kJ/kg, are
zero-moisture air at 0 °C.
}

