\name{h_t_w}
\alias{h_t_w}
\title{h_t_w}
\usage{
h_t_w(t, w)
}
\arguments{
  \item{t}{numeric: temperaure (°C)}

  \item{w}{numeric: humidity ratio (1)}
}
\value{
numeric: enthalpy of dry air (kJ/kg)
}
\description{
Enthalpy of moist air as function of temperature and
humidity ratio
}
\details{
The reference conditions, where enthalpy is 0 kJ/kg, are
zero-moisture air at 0 °C.
}

