\name{p_ws_t}
\alias{p_ws_t}
\title{p_ws_t}
\usage{
p_ws_t(t)
}
\arguments{
  \item{t}{numeric: temperature, -100 <= t <= 200 (°C)}
}
\value{
numeric: saturation pressure of water (Pa)
}
\description{
Saturation pressure of water in air as function of
temperature (°C)
}
\details{
The saturation pressure of water is the pressure at or
above which water would condense from from the air. It is a
function of temperature only.

For this function, valid temperature inputs are between
-100 and 200 °C.

Reference: 2005 ASHRAE Handbook - Fundamentals ch.6 (5),
(6)
}

