% Generated by roxygen2 (4.0.1): do not edit by hand
\name{t_wb}
\alias{t_wb}
\title{t_wb}
\usage{
t_wb(t, rh, p_atm = 101325)
}
\arguments{
\item{t}{numeric temperature (°C)}

\item{rh}{numeric relative humidity (1)}

\item{p_atm}{numeric atmospheric pressure (Pa)}
}
\value{
numeric wet-bulb temperature (°C)
}
\description{
given a set of temperatures, relative humidities, pressures,
determine wet-bulb temperatures
}

